refactor(rust): split academic http client domains
